User handler

+++ api/cmd/server/handler/user.go
@@ -36,8 +36,14 @@ func (u *User) AddAddress() gin.HandlerFunc {
 			c.JSON(400, web.NewResponse(400, nil, "Request body could not be binded"))
 			return
 		}
-		address, err := u.userService.AddAddress(ctx, address_input, user_id)
 
+		address, err := u.userService.AddAddress(ctx, address_input, user_id)
+		if err != nil {
+			c.JSON(400, web.NewResponse(400, nil, err.Error()))
+			return
+		}
+		c.JSON(200, web.NewResponse(201, address, ""))
+		return
 	}
 }
 
@@ -52,8 +58,19 @@ func (u *User) RemoveAddress() gin.HandlerFunc {
 		}
 
 		var address_input dto.RemoveAddress_Dto
+		bindingErr := c.ShouldBindJSON(&address_input)
+		if bindingErr != nil {
+			c.JSON(400, web.NewResponse(400, nil, "Request body could not be binded"))
+			return
+		}
 
 		addressId, err := u.userService.RemoveAddress(ctx, address_input, user_id)
+		if err != nil {
+			c.JSON(400, web.NewResponse(400, nil, err.Error()))
+			return
+		}
+		c.JSON(200, web.NewResponse(200, addressId, ""))
+		return
 	}
 }
 
@@ -68,9 +85,19 @@ func (u *User) AddCreditCard() gin.HandlerFunc {
 		}
 
 		var credit_card_input dto.InputCreditCard_Dto
+		bindingErr := c.ShouldBindJSON(&credit_card_input)
+		if bindingErr != nil {
+			c.JSON(400, web.NewResponse(400, nil, "Request body could not be binded"))
+			return
+		}
 
 		displayCreditCard, err := u.userService.AddCreditCard(ctx, credit_card_input, user_id)
-
+		if err != nil {
+			c.JSON(400, web.NewResponse(400, nil, err.Error()))
+			return
+		}
+		c.JSON(200, web.NewResponse(201, displayCreditCard, ""))
+		return
 	}
 }
 
@@ -85,8 +112,19 @@ func (u *User) RemoveCreditCard() gin.HandlerFunc {
 		}
 
 		var credit_card_input dto.RemoveCreditCard_Dto
+		bindingErr := c.ShouldBindJSON(&credit_card_input)
+		if bindingErr != nil {
+			c.JSON(400, web.NewResponse(400, nil, "Request body could not be binded"))
+			return
+		}
 
 		creditCardId, err := u.userService.RemoveCreditCard(ctx, credit_card_input, user_id)
+		if err != nil {
+			c.JSON(400, web.NewResponse(400, nil, err.Error()))
+			return
+		}
+		c.JSON(200, web.NewResponse(200, creditCardId, ""))
+		return
 	}
 }
 
@@ -100,9 +138,18 @@ func (u *User) ToggleProductWhishlist() gin.HandlerFunc {
 			c.JSON(400, web.NewResponse(400, nil, "User id could not be found"))
 		}
 
-		var product_id string
+		product_id := c.Param("product_id")
+		if product_id == "" {
+			c.JSON(400, web.NewResponse(400, nil, "Product id could not be found"))
+		}
 
 		err := u.userService.ToggleProductWhishlist(ctx, user_id, product_id)
+		if err != nil {
+			c.JSON(400, web.NewResponse(400, nil, err.Error()))
+			return
+		}
+		c.JSON(200, web.NewResponse(200, product_id, ""))
+		return
 	}
 }
 
@@ -117,6 +164,11 @@ func (u *User) AddProductToCart() gin.HandlerFunc {
 		}
 
 		var order dto.Order_Dto
+		bindingErr := c.ShouldBindJSON(&order)
+		if bindingErr != nil {
+			c.JSON(400, web.NewResponse(400, nil, "Request body could not be binded"))
+			return
+		}
 
 		orderId, err := u.userService.AddProductToCart(ctx, user_id, order.StockId, order.Quantity, order.Price)
 		if err != nil {
@@ -132,8 +184,16 @@ func (u *User) AddProductToCart() gin.HandlerFunc {
 func (u *User) RemoveProductFromCart() gin.HandlerFunc {
 	return func(c *gin.Context) {
 		ctx := context.Background()
-		// var user_id string
-		var order_id string
+
+		user_id := c.Request.Header.Get("user_id")
+		if user_id == "" {
+			c.JSON(400, web.NewResponse(400, nil, "User id could not be found"))
+		}
+
+		order_id := c.Param("order_id")
+		if order_id == "" {
+			c.JSON(400, web.NewResponse(400, nil, "Product id could not be found"))
+		}
 
 		err := u.userService.RemoveProductFromCart(ctx, order_id)
 		if err != nil {
@@ -149,8 +209,16 @@ func (u *User) RemoveProductFromCart() gin.HandlerFunc {
 func (u *User) IncreaseProductInCart() gin.HandlerFunc {
 	return func(c *gin.Context) {
 		ctx := context.Background()
-		// var user_id string
-		var order_id string
+
+		user_id := c.Request.Header.Get("user_id")
+		if user_id == "" {
+			c.JSON(400, web.NewResponse(400, nil, "User id could not be found"))
+		}
+
+		order_id := c.Param("order_id")
+		if order_id == "" {
+			c.JSON(400, web.NewResponse(400, nil, "Product id could not be found"))
+		}
 
 		err := u.userService.IncreaseProductInCart(ctx, order_id)
 		if err != nil {
@@ -166,8 +234,16 @@ func (u *User) IncreaseProductInCart() gin.HandlerFunc {
 func (u *User) DecreaseProductInCart() gin.HandlerFunc {
 	return func(c *gin.Context) {
 		ctx := context.Background()
-		// var user_id string
-		var order_id string
+
+		user_id := c.Request.Header.Get("user_id")
+		if user_id == "" {
+			c.JSON(400, web.NewResponse(400, nil, "User id could not be found"))
+		}
+
+		order_id := c.Param("order_id")
+		if order_id == "" {
+			c.JSON(400, web.NewResponse(400, nil, "Product id could not be found"))
+		}
 
 		err := u.userService.DecreaseProductInCart(ctx, order_id)
 		if err != nil {
